==> 02_Data Visualization WITH and WITHOUT Null Values.ipynb <==
# Data Visualisation WITH and WITHOUT Null Values

### Pandas Profiling
  - A one-line code EDA tool which generates interactive reports in a web format and requires no programming background to understand.
  
Advantages of this format is the features of each variable is shown such as Warnings, Number & Percentage of NaN Values, Cardinality & Skewness. 

import pandas as pd 
import numpy as np 
from pandas_profiling import ProfileReport

data_with_null = pd.read_csv(r"C:\Users\sak\Desktop\excelr_proj_1\data (1).csv",encoding='latin1')

data_without_null = pd.read_csv(r"C:\Users\sak\Desktop\excelr_proj_1\data_no_null.csv",encoding='latin1')

#profile = ProfileReport(df, title="Pandas Profiling Report")

profile_for_data_with_null=ProfileReport(data_with_null , title="Pandas Profiling Report For Data Having Null ")

#profile.to_file("your_report.html")
profile_for_data_with_null.to_file("report_with_null.html")

#profile = ProfileReport(df, title="Pandas Profiling Report")

profile_for_data_without_null=ProfileReport(data_without_null , title="Pandas Profiling Report For Data Having No Null ")

#profile.to_file("your_report.html")
profile_for_data_without_null.to_file("report_without_null.html")
