==> 04_Final Model.ipynb <==
# Final model

import pandas as pd 
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import r2_score as rsq
import xgboost as xgb
from feature_engine.categorical_encoders import OrdinalCategoricalEncoder
from feature_engine.categorical_encoders import CountFrequencyCategoricalEncoder
from sklearn.model_selection import RandomizedSearchCV
from sklearn.metrics import mean_squared_error

# reading the data(this data set not conataing null becoz we have imputed null vallues using cluster imputation)
data=pd.read_csv(r"C:\Users\sak\Desktop\excelr_proj_1\data_no_null.csv",encoding='latin1')# reading the data

# list of columns not necessary for prediction
cols_to_drop=['Unnamed: 0','sub_grade','State','Emp_designation','last_week_pay']
# dropping the unnecessary columns
df_xg=data.drop(columns=cols_to_drop,axis=1)

#df_xg

categorical = [col for col in df_xg.columns if df_xg[col].dtypes == 'O']

categorical

df_xg[['purpose']] = df_xg[['purpose']].replace(['car','house','renewable_energy','wedding','vacation','moving','medical','educational'],
                                              ['major_purchase','major_purchase','small_business','other','other','other','medical(or)education','medical(or)education'])

df_xg.purpose.value_counts().sort_values()

ordinal_enc = CountFrequencyCategoricalEncoder(
    encoding_method='frequency',
    variables=categorical)

df_xg_fre_enc = ordinal_enc.fit_transform(df_xg)

X_xg_f=df_xg_fre_enc[['loan_amnt ', 'terms', 'Rate_of_intrst', 'grade', 'home_ownership',
       'annual_inc', 'verification_status', 'purpose', 'debt_income_ratio',
       'delinq_2yrs', 'inq_last_6mths', 'numb_credit', 'pub_rec',
       'total_credits', 'initial_list_status',
       'total_rec_int', 'total_rec_late_fee', 'recoveries',
       'collection_recovery_fee', 'collections_12_mths_ex_med',
       'application_type', 'acc_now_delinq', 'Experience',
       'mths_since_last_delinq', 'tot_curr_bal', 'tot_colle_amt']]
Y_xg_f=df_xg_fre_enc[['total revol_bal']]

# let's separate into training and testing set

X_train_xg_f, X_test_xg_f, y_train_xg_f, y_test_xg_f = train_test_split(X_xg_f  ,  # predictors
                                                    Y_xg_f,  # target
    test_size=0.2,  # percentage of obs in test set
    random_state=2)  # seed to ensure reproducibility

X_train_xg_f.shape, X_test_xg_f.shape

xgb_mod_f=xgb.XGBRegressor()
params_f={
        'learning_rate':[0.03,0.05,0.08,0.10,0.15,0.20,0.25,0.30],
        'max_depth':[3,4,5,6,8,10,12,15,20,25],
        'min_child_weight':[1,3,5,7],
        'gamma':[0.0,0.1,0.2,0.3,0.4]
        }
random_search_f=RandomizedSearchCV(xgb_mod_f,param_distributions=params_f,n_iter=5,n_jobs=-1,cv=5,verbose=3)
random_search_f.fit(X_train_xg_f,y_train_xg_f)
random_search_f.best_estimator_
random_search_f.best_params_

xgb_mod_f=xgb.XGBRegressor(min_child_weight=5, max_depth=8, learning_rate=0.05, gamma=0.2)
xgb_mod_f.fit(X_train_xg_f,y_train_xg_f)


xg_pred_train_f=xgb_mod_f.predict(X_train_xg_f) 
xg_train_f_r2=rsq(y_train_xg_f,xg_pred_train_f)            
xg_train_f_r2

xg_pred_test_f=xgb_mod_f.predict(X_test_xg_f)
xg_test_f_r2=rsq(y_test_xg_f,xg_pred_test_f)
xg_test_f_r2

print("train R^2 :", xg_train_f_r2)
print("test R^2 :", xg_test_f_r2)

import sklearn.metrics as metrics
from sklearn.metrics import mean_squared_error
print(' train RMSE :', np.sqrt(metrics.mean_squared_error(y_train_xg_f,xg_pred_train_f)))
print(' test RMSE :', np.sqrt(metrics.mean_squared_error(y_test_xg_f,xg_pred_test_f)))

k_xg_f=list(xgb_mod_f.feature_importances_)
j_xg_f=list(X_train_xg_f.columns)
data_xg_f ={'columns': j_xg_f, 'feature importance':k_xg_f}
d_xg_f=pd.DataFrame(data_xg_f)
d_xg_f

#### so here i m droping the features whose importance is less than 1% :-
collection_recovery_fee

collections_12_mths_ex_med

application_type

acc_now_delinq

X_xg_f_1=df_xg_fre_enc[['loan_amnt ', 'terms', 'Rate_of_intrst', 'grade', 'home_ownership','annual_inc', 'verification_status',
                        'purpose', 'debt_income_ratio','delinq_2yrs', 'inq_last_6mths', 'numb_credit', 'pub_rec', 'total_credits', 
                        'initial_list_status','total_rec_int', 'total_rec_late_fee', 'recoveries','Experience','mths_since_last_delinq', 
                        'tot_curr_bal', 'tot_colle_amt']]
Y_xg_f_1=df_xg_fre_enc[['total revol_bal']]

# let's separate into training and testing set

X_train_xg_f_1, X_test_xg_f_1, y_train_xg_f_1, y_test_xg_f_1 = train_test_split(X_xg_f_1  ,  # predictors
                                                    Y_xg_f_1,  # target
    test_size=0.2,  # percentage of obs in test set
    random_state=2)  # seed to ensure reproducibility

X_train_xg_f_1.shape, X_test_xg_f_1.shape

xgb_mod_f_1=xgb.XGBRegressor(min_child_weight=5, max_depth=8, learning_rate=0.05, gamma=0.2)
xgb_mod_f_1.fit(np.array(X_train_xg_f_1),np.array(y_train_xg_f_1))


xg_pred_train_f_1=xgb_mod_f_1.predict(np.array(X_train_xg_f_1)) 
xg_train_f_1_r2=rsq(y_train_xg_f_1,xg_pred_train_f_1)            

xg_pred_test_f_1=xgb_mod_f_1.predict(np.array(X_test_xg_f_1)) 
xg_test_f_1_r2=rsq(y_test_xg_f_1,xg_pred_test_f_1)            

print("train R^2 :", xg_train_f_1_r2)
print("test R^2 :", xg_test_f_1_r2)

import sklearn.metrics as metrics
from sklearn.metrics import mean_squared_error
print(' train RMSE :', np.sqrt(metrics.mean_squared_error(y_train_xg_f_1,xg_pred_train_f_1)))
print(' test RMSE :', np.sqrt(metrics.mean_squared_error(y_test_xg_f_1,xg_pred_test_f_1)))

k_xg_f_1=list(xgb_mod_f_1.feature_importances_)
j_xg_f_1=list(X_train_xg_f_1.columns)
data_xg_f_1 ={'columns': j_xg_f_1, 'feature importance':k_xg_f_1}
d_xg_f_1=pd.DataFrame(data_xg_f_1)
d_xg_f_1

import pickle
pickle.dump(xgb_mod_f_1,open('finalized_model.pkl','wb'))

X_test_xg_f_1

# prediction using the saved model.
loaded_model = pickle.load(open('finalized_model.pkl', 'rb'))
prediction=loaded_model.predict(np.array(X_test_xg_f_1))

print(prediction)


##### so above one is showing our saved model is working fine 

### Saving frequency-encoding of "terms" to disk ----- later used in the Deployment

a = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['terms'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['terms']].values
temp_values = a.fit_transform(df_xg[['terms']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['terms'] = dict_temp
import pickle
filehandler = open("terms.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("terms.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

### Saving frequency-encoding of "grade" to disk ----- later used in the Deployment

b = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['grade'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['grade']].values
temp_values = b.fit_transform(df_xg[['grade']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['grade'] = dict_temp
import pickle
filehandler = open("grade.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("grade.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

### Saving frequency-encoding of "home_ownership" to disk ----- later used in the Deployment

c = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['home_ownership'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['home_ownership']].values
temp_values = c.fit_transform(df_xg[['home_ownership']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['home_ownership'] = dict_temp
import pickle
filehandler = open("home_ownership.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("home_ownership.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

df_xg.home_ownership.value_counts().sort_index()

### Saving frequency-encoding of "verification_status" to disk ----- later used in the Deployment

d = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['verification_status'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['verification_status']].values
temp_values = d.fit_transform(df_xg[['verification_status']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['verification_status'] = dict_temp
import pickle
filehandler = open("verification_status.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("verification_status.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

df_xg.verification_status.value_counts().sort_index()

### Saving frequency-encoding of "purpose" to disk ----- later used in the Deployment

e = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['purpose'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['purpose']].values
temp_values = e.fit_transform(df_xg[['purpose']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['purpose'] = dict_temp
import pickle
filehandler = open("purpose.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("purpose.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

df_xg.purpose.value_counts().sort_index()

### Saving frequency-encoding of "initial_list_status" to disk ----- later used in the Deployment

f = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['initial_list_status'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['initial_list_status']].values
temp_values = f.fit_transform(df_xg[['initial_list_status']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['initial_list_status'] = dict_temp
import pickle
filehandler = open("initial_list_status.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("initial_list_status.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

df_xg.initial_list_status.value_counts().sort_index()

### Saving frequency-encoding of "Experience" to disk ----- later used in the Deployment

g = CountFrequencyCategoricalEncoder(encoding_method='frequency',variables=['Experience'])
dict_all = dict(zip([], []))
temp_keys = df_xg[['Experience']].values
temp_values = g.fit_transform(df_xg[['Experience']])
dict_temp = dict(zip(np.array(temp_keys).ravel(),np.array(temp_values).ravel()))
dict_all['Experience'] = dict_temp
import pickle
filehandler = open("Experience.obj","wb")
pickle.dump(dict_all,filehandler)
filehandler.close()

file = open("Experience.obj",'rb')
terms = pickle.load(file)
file.close()
print(terms)

df_xg.Experience.value_counts().sort_index()
