==> 05_Model Testing.ipynb <==
import pickle
import numpy as np
from flask import Flask, request, render_template
import xgboost as xgb

terms = pickle.load(open('terms.obj','rb'))
grade = pickle.load(open('grade.obj','rb'))
home_ownership = pickle.load(open('home_ownership.obj','rb'))
verification_status = pickle.load(open('verification_status.obj','rb'))
purpose = pickle.load(open('purpose.obj','rb'))
initial_list_status = pickle.load(open('initial_list_status.obj','rb'))
Experience = pickle.load(open('Experience.obj','rb'))
model = pickle.load(open('finalized_model.pkl','rb'))

#Experience

    features=[]
    a= (float('14350'))
    features.append(a)
    b='36 months'
    features.append(terms['terms'][b])
    c=(float('19.19'))
    features.append(c)
    d='E'
    features.append(grade['grade'][d])
    e='OWN'
    features.append(home_ownership['home_ownership'][e])
    f=(float('28700'))
    features.append(f)
    g='Source Verified'
    features.append(verification_status['verification_status'][g])
    h='debt_consolidation'
    features.append(purpose['purpose'][h])
    i=(float('33.88'))
    features.append(i)
    j=(float('0'))
    features.append(j)
    k=(float('1'))
    features.append(k)
    l=(float('14'))
    features.append(l)
    m=(float('1'))
    features.append(m)
    n=(float('28'))
    features.append(n)
    o='f'
    features.append(initial_list_status['initial_list_status'][o])
    p=(float('1173.84'))
    features.append(p)
    q=(float('0'))
    features.append(q)
    r=(float('0'))
    features.append(r)
    s='9 years'
    features.append(Experience['Experience'][s])
    t=(float('50'))
    features.append(t)
    u=(float('28699'))
    features.append(u)
    v=(float('0'))
    features.append(v)

features

final_features = np.array(features).reshape(-1,22)
prediction = model.predict(final_features)
prediction
